fix error convert values cart

+++ src/main/java/com/example/phone_store/entity/Order.java
@@ -4,12 +4,11 @@ import java.util.Date;
 
 public class Order {
     private Integer orderId;
-    private Integer userId;
+    private Long userId;
     private Double totalAmount;
     private String status;
     private Date orderDate;
 
-    // Thông tin giao hàng
     private String receiverName;
     private String phone;
     private String shippingAddress;
@@ -18,8 +17,8 @@ public class Order {
     public Integer getOrderId() { return orderId; }
     public void setOrderId(Integer orderId) { this.orderId = orderId; }
 
-    public Integer getUserId() { return userId; }
-    public void setUserId(Integer userId) { this.userId = userId; }
+    public Long getUserId() { return userId; }
+    public void setUserId(Long userId) { this.userId = userId; }
 
     public Double getTotalAmount() { return totalAmount; }
     public void setTotalAmount(Double totalAmount) { this.totalAmount = totalAmount; }
